pyproject.toml
[tool.poetry]
name = "vfzsync"
version = "0.4.0"
description = ""
authors = ["Daniil Ovsyannikov <to.djet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
py-zabbix = "^1.1"
vpoller = "^0.7.3"
PyYAML = "^5.1"
python-dateutil = "^2.8.1"
flask = "^1.1.1"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
jupyterlab = "^1.2"
jupyter = "^1.0"
jupyter-server = "^0.1.1"
pandas = "^0.25.3"
rope = "^0.14.0"
black = "^19.10b0"
nose = "^1.3.7"
parametrized = "^0.1"
nose_parameterized = "^0.6.0"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
